+++ tests/workflows/mnist-fedavg/assets/aggregate_function.py
@@ -34,7 +34,7 @@ Function that aggregates models by simply averaging them as in FedAvg
 def aggregate(inputs, outputs, task_properties):
     # get layers
     inmodels = []
-    for m_path in inputs["models"]:
+    for m_path in inputs["shared"]:
         inmodels.append(load_model(m_path))
 
     model = inmodels[0]
@@ -50,7 +50,7 @@ def aggregate(inputs, outputs, task_properties):
 
     model.load_state_dict(model_state_dict)
 
-    save_model(model, outputs["model"])
+    save_model(model, outputs["shared"])
 
 
 @tools.register
@@ -58,7 +58,7 @@ def predict(inputs, outputs, task_properties):
     X = inputs["datasamples"]["X"]
     X = torch.FloatTensor(X)
 
-    model = load_model(inputs["models"])
+    model = load_model(inputs["shared"])
     model.eval()
     # add the context manager to reduce computation overhead
     with torch.no_grad():